🧹 Refactor - Cleanup + Refactor Functionality #8 - Refactor into functions
