+++ src/constants.ts
@@ -1,0 +1,11 @@
+import { Day } from "./types";
+
+export const days: Day[] = [
+  "Sunday",
+  "Monday",
+  "Tuesday",
+  "Wednesday",
+  "Thursday",
+  "Friday",
+  "Saturday",
+];
